+++ tests/geometryRTheta/advection_rtheta/advection_all_tests.cpp
@@ -33,7 +33,7 @@
 #include "rk2.hpp"
 #include "rk3.hpp"
 #include "rk4.hpp"
-#include "spline_interpolator_rtheta.hpp"
+#include "spline_interpolator_2d.hpp"
 #include "spline_polar_foot_finder.hpp"
 #include "test_cases.hpp"
 #include "vector_field.hpp"
