+++ tests/geometryRTheta/spline_interpolator_rtheta/spline_interpolator_rtheta.cpp
@@ -12,7 +12,7 @@
 #include "geometry.hpp"
 #include "paraconfpp.hpp"
 #include "params.yaml.hpp"
-#include "spline_interpolator_rtheta.hpp"
+#include "spline_interpolator_2d.hpp"
 
 
 /*
